requirement.txt
appdirs==1.4.4
asn1crypto==0.24.0
Babel==2.8.0
certifi==2018.1.18
chardet==3.0.4
countries==0.2.0
cryptography==2.1.4
defusedxml==0.6.0
distlib==0.3.0
Django==2.2
django-allauth==0.42.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.6
importlib-metadata==1.6.1
importlib-resources==1.5.0
keyring==10.6.0
keyrings.alt==3.0
#anguage-selector==0.1
Markdown==3.2.2
netifaces==0.10.4
oauthlib==3.1.0
Pillow==7.1.2
pycrypto==2.6.1
#pygobject==3.26.1
#python-apt==1.6.5+ubuntu0.3
python3-openid==3.1.0
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
requests==2.23.0
requests-oauthlib==1.3.0
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.3.1
#ish-import-id==5.7
stripe==2.48.0
#ufw==0.36
urllib3==1.22
virtualenv==20.0.21
#zipp==3.1.0
